+++ src/ui/Player/Clip/ClipClosed.css
